requirements.txt
certifi==2020.12.5
chardet==4.0.0
feedparser==6.0.2
gitterpy==0.1.7
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.15.0
PyYAML==5.4.1
requests==2.25.1
sgmllib3k==1.0.0
urllib3==1.26.3
